+++ tests/testthat/test-factorial.R
@@ -42,6 +42,6 @@ test_that("Factorial", {
 
   expect_equal(diagnosis %>% get_simulations %>% dim, c(2, 14))
 
-  expect_equal(diagnosis %>%  get_diagnosands %>% dim, c(1, 14))
+  expect_equal(diagnosis %>%  get_diagnosands %>% dim, c(1, 12))
 
 })